feature : 영수증 Loading 기능
영수증의 명언이 fetch하는 동안 영수증대신 로딩 컴포넌트를 보여줄 수 있도록 했습니다.
영수증 명언이 늦게 불러와지면서 일어나는 레이아웃 시프트와 사용자경험 저하를 막기 위해서 로딩을 추가하는 쪽이 낫다고 판단하여 개선하였습니다.
