+++ src/components/atoms/ReceiptQuotes/ReceiptQuotes.jsx
@@ -1,31 +1,12 @@
 import * as S from "./ReceiptQuotes.styles";
-import useAsync from "../../../hooks/useAsync";
-import { useEffect } from "react";
-
 /**
  * ReceiptQuotes
  *
  * 영수증의 문구를 보여주는 컴포넌트입니다
  * 랜덤 영어 명언 API를 사용해 랜덤으로 문구를 보여줍니다
- *
+ * @param {string} quote 명언
  * @returns
  */
-export function ReceiptQuotes() {
-  const [state] = useAsync(getQuotes, []);
-  const { loading, data: quote, error } = state;
-
-  // TODO: 리팩토링 예정
-  useEffect(() => {
-    sessionStorage.setItem("famous_saying", quote);
-  }, [quote]);
-
-  return <S.Quotes>{loading || error ? "Well done!" : quote}</S.Quotes>;
-}
-
-async function getQuotes() {
-  const response = await fetch("https://api.adviceslip.com/advice").then((response) =>
-    response.json(),
-  );
-
-  return response.slip.advice;
+export function ReceiptQuotes({ quote }) {
+  return <S.Quotes>{quote}</S.Quotes>;
 }
