+++ src/components/molecules/ReceiptPaper/ReceiptPaper.jsx
@@ -1,5 +1,8 @@
 import { ReceiptPaperTriangle, ReceiptPaperContents } from "components";
 import * as S from "./ReceiptPaper.styles";
+import useAsync from "../../../hooks/useAsync";
+import { useEffect } from "react";
+import Loading from "react-loading";
 
 /**
  * ReceiptPaper
@@ -13,17 +16,36 @@ import * as S from "./ReceiptPaper.styles";
  */
 
 export function ReceiptPaper({ todos, onClick = null }) {
+  const [state] = useAsync(getQuotes, []);
+  const { loading, data: quote, error } = state;
+
+  useEffect(() => {
+    sessionStorage.setItem("famous_saying", quote);
+  }, [quote]);
+
   return (
     <S.Container id={"receipt"} onClick={onClick}>
-      <div>
-        <S.TrianglePosition y="6px">
-          <ReceiptPaperTriangle isPaper={true} />
-        </S.TrianglePosition>
-        <ReceiptPaperContents todos={todos} />
-        <S.TrianglePosition y="-7px">
-          <ReceiptPaperTriangle isPaper={true} />
-        </S.TrianglePosition>
-      </div>
+      {loading ? (
+        <Loading />
+      ) : (
+        <div>
+          <S.TrianglePosition y="6px">
+            <ReceiptPaperTriangle isPaper={true} />
+          </S.TrianglePosition>
+          <ReceiptPaperContents todos={todos} quote={error ? "Well done!" : quote} />
+          <S.TrianglePosition y="-7px">
+            <ReceiptPaperTriangle isPaper={true} />
+          </S.TrianglePosition>
+        </div>
+      )}
     </S.Container>
   );
+}
+
+async function getQuotes() {
+  const response = await fetch("https://api.adviceslip.com/advice").then((response) =>
+    response.json(),
+  );
+
+  return response.slip.advice;
 }
