+++ src/components/molecules/ReceiptPaperContents/ReceiptPaperContents.jsx
@@ -7,9 +7,10 @@ import * as S from "./ReceiptPaperContents.styles";
  * 영수증의 내용이 들어가는 컴포넌트입니다
  *
  * @param {Array} todos todo 리스트
+ * @param {string} quote 명언
  * @returns
  */
-export function ReceiptPaperContents({ todos }) {
+export function ReceiptPaperContents({ todos, quote }) {
   const line = "-----------------------------------";
   let timerTotal = 0;
   return (
@@ -37,7 +38,7 @@ export function ReceiptPaperContents({ todos }) {
         </div>
       </S.TotalContainer>
       <div>{line}</div>
-      <ReceiptQuotes />
+      <ReceiptQuotes quote={quote} />
       <div>{line}</div>
       <ReceiptBarcode />
       <div>{line}</div>
